Add support for handling messages
The transaction pool has been replaced by a message pool. Clients can
submit messages to the node via the /api/v0/messages endpoint. The
/api/v0/transaction endpoint has been kept for compatibility. The
leadership task pulls messages from the message pool to create blocks.

As an example, I added a command to jormungandr to create a stake
pool. This should probably be moved to cardano-cli eventually. For example:

  $ cargo run create-stake-pool --host http://localhost:8080/
  PrivateKey(f863fc82e18109bfe23e61ff8a0c05c29eeaf7effe971edc0c1b22bb8fae4147)
  PublicKey(ff79936ed09cc4076b9f4549fd7c3c3934083b67e2ecda4f5e0c26515e263a7b)

We can check whether the stake pool was created succesfully:

  $ curl http://localhost:8080/api/v0/pool
  {"StakePoolId(PublicKey(ff79936ed09cc4076b9f4549fd7c3c3934083b67e2ecda4f5e0c26515e263a7b))":{"members":[]}}